<Player skill manager access for affliction bonues> [update]
+ add access on player skil lmanager for affliction bonuses

+++ Assets/Scripts/Scriptable Object Scripts/Player Skill Data/AfflictionUpgradeData.cs
@@ -52,7 +52,7 @@ public class AfflictionUpgradeData
         return rateChance[p_currentLevel];
     }
 
-    public float GetCrowdNumberPerLevel(int p_currentLevel) {
+    public int GetCrowdNumberPerLevel(int p_currentLevel) {
         if (crowdNumber == null || crowdNumber.Count <= 0) {
             return -1;
         }
